requirements.txt
Flask==0.10.1
PyYAML==3.11
requests==2.5.3
slacker==0.5.7